requirements.txt
openai
llama-index
flask
pydantic
sentence-transformers
transformers
python-dotenv
